Validation and Error Handling Added

+++ user/user.go
@@ -22,7 +22,7 @@ type User struct {
 	Username string `json:"username"`
 	Password string `json:"password"`
 	Name     string `json:"name"`
-	Role     string `json:"role"`
+	Role     string `json:"role" gorm:"default:user"`
 }
 
 func GetUsers(c *fiber.Ctx) error {
@@ -39,6 +39,15 @@ func Authenticate(c *fiber.Ctx) error {
 	err := fiberutils.ParseBody(c, user) //(c (pagkukuhanan), user (pagpapasahan))
 	username := user.Username
 	password := user.Password
+
+	if len(username) == 0 || len(password) == 0 {
+		c.JSON(fiberutils.Message{
+			Message: "Please Input Username and Password",
+			Status:  "error",
+		})
+		return nil
+	}
+
 	db.First(user, "username = ?", user.Username) //(user , "username = ?" (?-placeholder), user.Username)
 
 	if passwordhashing.CheckPasswordHash(password, user.Password) {
@@ -53,7 +62,7 @@ func Authenticate(c *fiber.Ctx) error {
 		// Set claims
 		claims := token.Claims.(jwt.MapClaims)
 		claims["user"] = user
-		claims["exp"] = time.Now().Add(time.Minute * 15).Unix()
+		claims["exp"] = time.Now().Add(time.Hour * 72).Unix()
 
 		if user.Role == "admin" {
 			claims["admin"] = true
@@ -98,6 +107,34 @@ func NewUser(c *fiber.Ctx) error {
 	db := database.DBConn
 	user := new(User)
 	err := fiberutils.ParseBody(c, user)
+	var usernameCount int64
+	db.Model(&user).Where("username = ?", user.Username).Count(&usernameCount)
+	usernameExists := usernameCount > 0
+
+	if len(user.Username) == 0 || len(user.Password) == 0 || len(user.Name) == 0 {
+		c.JSON(fiberutils.Message{
+			Message: "Please Input Username, Password and Name",
+			Status:  "error",
+		})
+		return nil
+	}
+
+	if len(user.Username) < 3 || len(user.Password) < 8 || len(user.Name) < 3 {
+		c.JSON(fiberutils.Message{
+			Message: "Required Mininum Length of Username, Name and Password is 3, 3 and 8 respectively",
+			Status:  "error",
+		})
+		return nil
+	}
+
+	if usernameExists {
+		c.JSON(fiberutils.Message{
+			Message: "Username Already Exists",
+			Status:  "error",
+		})
+		return nil
+	}
+
 	user.Password, err = passwordhashing.HashPassword(user.Password)
 	db.Create(&user) //db. is from gorm package which automatically create a query statement either create, update, delete or view //&-get reference from user
 	c.JSON(fiberutils.Message{
@@ -117,8 +154,34 @@ func UpdateUser(c *fiber.Ctx) error {
 	err := fiberutils.ParseBody(c, user)
 
 	if userO["role"] == "admin" || fmt.Sprintf("%v", userO["id"]) == fmt.Sprintf("%d", user.Id) {
+		if len(user.Username) == 0 || len(user.Password) == 0 || len(user.Name) == 0 {
+			c.JSON(fiberutils.Message{
+				Message: "Please Input Username, Password and Name",
+				Status:  "error",
+			})
+			return nil
+		}
+
+		if len(user.Username) < 3 || len(user.Password) < 8 || len(user.Name) < 3 {
+			c.JSON(fiberutils.Message{
+				Message: "Required Mininum Length of Username, Name and Password is 3, 3 and 8 respectively",
+				Status:  "error",
+			})
+			return nil
+		}
+
 		user.Password, err = passwordhashing.HashPassword(user.Password)
-		db.Model(&user).Where("id = ?", user.Id).Updates(user)
+		db.First(&user, user.Id)
+
+		if len(user.Role) == 0 {
+			c.JSON(fiberutils.Message{
+				Message: fmt.Sprintf("No User with Id: %d", user.Id),
+				Status:  "error",
+			})
+			return nil
+		}
+
+		db.Updates(&user)
 		c.JSON(fiberutils.Message{
 			Message: "User Successfully Updated",
 			Status:  "success",
@@ -142,6 +205,14 @@ func DeleteUser(c *fiber.Ctx) error {
 	id := c.Params("id")
 	user := new(User)
 	db.First(&user, id)
+
+	if len(user.Role) == 0 {
+		c.JSON(fiberutils.Message{
+			Message: "No User with Id: " + id,
+			Status:  "error",
+		})
+		return nil
+	}
 
 	if userO["role"] == "admin" || fmt.Sprintf("%v", userO["id"]) == fmt.Sprintf("%d", user.Id) {
 		db.Delete(&user)
